fix build - can't have non page files in pages.
